refactor: Consolidate ProcessSingleSession test into eventserver_test.go
