+++ historyserver/pkg/eventserver/eventserver_test.go
@@ -9,6 +9,8 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/ray-project/kuberay/historyserver/pkg/eventserver/types"
 	"github.com/ray-project/kuberay/historyserver/pkg/utils"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func makeTaskEventMap(taskName, nodeId, taskID, cluster string, attempt int) map[string]any {
@@ -994,4 +996,89 @@ func TestNormalizeActorIDsToHex(t *testing.T) {
 	if actor.Address.WorkerID != expectedHex {
 		t.Errorf("Address.WorkerID = %q, want %q", actor.Address.WorkerID, expectedHex)
 	}
+}
+
+func TestProcessSingleSession(t *testing.T) {
+	clusterInfo := utils.ClusterInfo{Name: "cluster", Namespace: "ns", SessionName: "session1"}
+
+	t.Run("returns error when every listed file fails I/O", func(t *testing.T) {
+		mock := newLogEventMockReader()
+		mock.addDir("cluster_ns", "session1/job_events/", []string{"job-01000000/"})
+		mock.addDir("cluster_ns", "session1/job_events/job-01000000/",
+			[]string{"01000000-2024-01-01-00"})
+		mock.addDir("cluster_ns", "session1/node_events/",
+			[]string{"node1-2024-01-01-00"})
+		mock.addDir("cluster_ns", "session1/logs", []string{})
+
+		h := NewEventHandler(mock)
+		err := h.ProcessSingleSession(context.Background(), clusterInfo)
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "ingested 0 of 2")
+	})
+
+	t.Run("empty file list returns nil (legit empty session)", func(t *testing.T) {
+		mock := newLogEventMockReader()
+		mock.addDir("cluster_ns", "session1/job_events/", []string{})
+		mock.addDir("cluster_ns", "session1/node_events/", []string{})
+		mock.addDir("cluster_ns", "session1/logs", []string{})
+
+		h := NewEventHandler(mock)
+		err := h.ProcessSingleSession(context.Background(), clusterInfo)
+		require.NoError(t, err)
+	})
+
+	t.Run("partial success does not return error", func(t *testing.T) {
+		mock := newLogEventMockReader()
+		mock.addDir("cluster_ns", "session1/node_events/",
+			[]string{"node1-2024-01-01-00", "node2-2024-01-01-00"})
+		mock.addFile("cluster_ns", "session1/node_events/node1-2024-01-01-00",
+			`[{"eventType":"NODE_DEFINITION_EVENT","nodeDefinitionEvent":{"nodeId":"YWJjZA=="}}]`)
+		mock.addDir("cluster_ns", "session1/job_events/", []string{})
+		mock.addDir("cluster_ns", "session1/logs", []string{})
+
+		h := NewEventHandler(mock)
+		err := h.ProcessSingleSession(context.Background(), clusterInfo)
+		require.NoError(t, err)
+	})
+
+	t.Run("all corrupt JSON does not return error", func(t *testing.T) {
+		mock := newLogEventMockReader()
+		mock.addDir("cluster_ns", "session1/node_events/",
+			[]string{"node1-2024-01-01-00", "node2-2024-01-01-00"})
+		mock.addFile("cluster_ns", "session1/node_events/node1-2024-01-01-00", "this is not json")
+		mock.addFile("cluster_ns", "session1/node_events/node2-2024-01-01-00", "neither is this")
+		mock.addDir("cluster_ns", "session1/job_events/", []string{})
+		mock.addDir("cluster_ns", "session1/logs", []string{})
+
+		h := NewEventHandler(mock)
+		err := h.ProcessSingleSession(context.Background(), clusterInfo)
+		require.NoError(t, err)
+	})
+
+	t.Run("ReadLogEvents error propagates even when RayEvents path is empty", func(t *testing.T) {
+		mock := newLogEventMockReader()
+		mock.addDir("cluster_ns", "session1/job_events/", []string{})
+		mock.addDir("cluster_ns", "session1/node_events/", []string{})
+		mock.addDir("cluster_ns", "session1/logs", []string{"node1/"})
+		mock.addDir("cluster_ns", "session1/logs/node1/events", []string{"event_GCS.log"})
+
+		h := NewEventHandler(mock)
+		err := h.ProcessSingleSession(context.Background(), clusterInfo)
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "ingested 0 of 1 log event files")
+	})
+
+	t.Run("both subsystems failing yields combined error", func(t *testing.T) {
+		mock := newLogEventMockReader()
+		mock.addDir("cluster_ns", "session1/node_events/", []string{"node1-2024-01-01-00"})
+		mock.addDir("cluster_ns", "session1/job_events/", []string{})
+		mock.addDir("cluster_ns", "session1/logs", []string{"node1/"})
+		mock.addDir("cluster_ns", "session1/logs/node1/events", []string{"event_GCS.log"})
+
+		h := NewEventHandler(mock)
+		err := h.ProcessSingleSession(context.Background(), clusterInfo)
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "ingested 0 of 1 RayEvent files")
+		assert.Contains(t, err.Error(), "ingested 0 of 1 log event files")
+	})
 }
